Game APK File
The APK File can be downloaded directly, if u cant run the app in android studio. 
Note: By downloading this APK file, u wont be able to see the contents of the Database because it's contents is logged directly into the logging console of android studio. U can only clear the database.